Split inputs logic into multiple files
